Port 5475 ocean kill the integration if no partition assignment was made for the consumer (#392)
Description
What - If there are two instances sharing the same identifier for the integration type with a Kafka event listener, the second instance will remain outdated and unassigned to any partitions. Therefor, it will not get any resyncs. We want to terminate the redundant instance since Ocean does not support multiple instances concurrently.
Why - Ocean does not support multiple instances for the same integration & the kafka topic has only one partition.
How - Checking if the integration received any partitions and shutting down the integration if not

Type of change
 New feature (non-breaking change which adds functionality)