[Integration][Kubecost] - Support backwards compatibility (#631)
# Description

What - The previous version of the integration broke the backwards
compatibility checks. This PR fixes it by adding a configurable
parameter `KUBECOST_API_VERSION`, that allows users to specify which
version of Kubecost they are using. If users are using the old version
(v1), we make a call to `/model/cloudCost/aggregate` and if they are
using v2, we call `/model/cloudCost`

Why -
How -

## Type of change

Please leave one option from the following and delete the rest:

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] New Integration (non-breaking change which adds a new integration)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Non-breaking change (fix of existing functionality that will not
change current behavior)
- [ ] Documentation (added/updated documentation)

## Screenshots

Include screenshots from your environment showing how the resources of
the integration will look.

## API Documentation

Provide links to the API documentation used for this integration.