[Integration][ADO] Set the `isProjectsLimited` paramater to True by default (#806)
# Description

Fixed [a bug in the Azure DevOps (ADO)
integration](https://port-community.slack.com/archives/C05LZB72NM8/p1718010461875529)
where project-scoped webhooks were not being created if the
isProjectsLimited parameter was not explicitly set. The default value
for this parameter is now set to True, ensuring that webhooks are
created with the correct scope.

## Type of change

Please leave one option from the following and delete the rest:

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] New Integration (non-breaking change which adds a new integration)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Non-breaking change (fix of existing functionality that will not
change current behavior)
- [x] Documentation (added/updated documentation)

## Screenshots

Include screenshots from your environment showing how the resources of
the integration will look.

## API Documentation

Provide links to the API documentation used for this integration.