[Integration][Servicenow] - Update Connectivity Test (#825)
# Description

What - Updated the Servicenow connectivity test to use an existing table
(incident) rather than the instance table, which may require the user to
grant extra permission
Why - As part of the connectivity check, we query the
tapi/now/table/instance endpoint to retrieve information about the
instance. One customer was particularly worried that they don't want the
integration to access this table since it involve granting admin
privileges. So, instead of querying another table, we use one of the
existing tables which we have access to(incident) to perform the
connectivity check. For us to know that the connection was successful,
we display a sample incident to the console

How -

## Type of change

Please leave one option from the following and delete the rest:

- [ ] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] New Integration (non-breaking change which adds a new integration)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Non-breaking change (fix of existing functionality that will not
change current behavior)
- [ ] Documentation (added/updated documentation)

## Screenshots

Include screenshots from your environment showing how the resources of
the integration will look.
<img width="1028" alt="Screenshot 2024-07-17 at 6 26 55 PM"
src="https://github.com/user-attachments/assets/f5843575-d13e-4da6-a589-390834b76b64">


## API Documentation

Provide links to the API documentation used for this integration.