[Gitlab] Improvements to file kind (#942)
# Description

What - Added improvements in resync + realtime analysis for file-kind.
Why - Have better scale, performance and verbosity for the feature
How - In resync- add a search in all repositories being processed,
making sure the path we need is present in the repo, and only then the
integration will pull the file. In real-time - added pagination for
verbosity

## Type of change

Please leave one option from the following and delete the rest:

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] New Integration (non-breaking change which adds a new integration)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Non-breaking change (fix of existing functionality that will not
change current behavior)
- [ ] Documentation (added/updated documentation)

## Screenshots

Include screenshots from your environment showing how the resources of
the integration will look.

## API Documentation

Provide links to the API documentation used for this integration.