[Integration][GCP] Added an alternative for retrieving quota project id (#997)
# Description

**What**  
This PR fixes a bug related to fetching the GCP project ID from service
account credentials. Previously, the function would fail to retrieve the
correct project ID when the `quota_project_id` key was missing from the
credentials file.

**Why**  
GCP configuration differs based on how the credentials are generated.
When using `gcloud auth application-default login`, the project ID is
stored under the `quota_project_id` key. However, when a service account
key is generated from the console, the project ID is stored under the
`project_id` key. This discrepancy caused the integration to fail in
identifying the project ID when the credentials were generated from the
console, impacting quota computation operations.

This bug affected users by causing failures in environments where
`quota_project_id` was not available. The fix ensures that both
configurations are supported, making the application more robust and
reliable in different environments.

**How**  
- Introduced a mechanism to check for both `project_id` and
`quota_project_id` keys in the credentials file.
- Improved error handling to raise meaningful exceptions when
credentials are missing or incomplete.


## Type of change

Please leave one option from the following and delete the rest:

- [x] Bug fix (non-breaking change which fixes an issue)
- [ ] New feature (non-breaking change which adds functionality)
- [ ] New Integration (non-breaking change which adds a new integration)
- [ ] Breaking change (fix or feature that would cause existing
functionality to not work as expected)
- [ ] Non-breaking change (fix of existing functionality that will not
change current behavior)
- [ ] Documentation (added/updated documentation)

<h4> All tests should be run against the port production
environment(using a testing org). </h4>


### Integration testing checklist

- [x] Integration able to create all default resources from scratch
- [x] Resync able to create entities
- [x] Resync able to update entities
- [x] Resync able to detect and delete entities
- [x] Resync finishes successfully
- [x] If new resource kind is added or updated in the integration, add
example raw data, mapping and expected result to the `examples` folder
in the integration directory.
- [x] If resource kind is updated, run the integration with the example
data and check if the expected result is achieved
- [x] If new resource kind is added or updated, validate that
live-events for that resource are working as expected
- [ ] Docs PR link [here](#)